refactor(MCP): move ModelScope sync to providers folder
Refactors ModelScope sync utilities to match the structure of other
providers for better organization and consistency.

Changes:
- Moved modelscopeSyncUtils.ts to providers/modelscope.ts
- Added MODELSCOPE_HOST constant for consistency
- Updated import statements in SyncServersPopup.tsx
- Maintained all existing functionality including updateServer support

Benefits:
- Consistent file organization across all sync providers
- Better maintainability with centralized provider structure
- Follows established patterns from other providers
