+++ src/renderer/src/pages/settings/MCPSettings/SyncServersPopup.tsx
@@ -6,9 +6,9 @@ import { useCallback, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 
-import { getModelScopeToken, saveModelScopeToken, syncModelScopeServers } from './modelscopeSyncUtils'
 import { getAI302Token, saveAI302Token, syncAi302Servers } from './providers/302ai'
 import { getTokenLanYunToken, LANYUN_KEY_HOST, saveTokenLanYunToken, syncTokenLanYunServers } from './providers/lanyun'
+import { getModelScopeToken, MODELSCOPE_HOST, saveModelScopeToken, syncModelScopeServers } from './providers/modelscope'
 import { getTokenFluxToken, saveTokenFluxToken, syncTokenFluxServers, TOKENFLUX_HOST } from './providers/tokenflux'
 
 // Provider configuration interface
@@ -30,8 +30,8 @@ const providers: ProviderConfig[] = [
     key: 'modelscope',
     name: 'ModelScope',
     description: 'ModelScope 平台 MCP 服务',
-    discoverUrl: 'https://www.modelscope.cn/mcp?hosted=1&page=1',
-    apiKeyUrl: 'https://www.modelscope.cn/my/myaccesstoken',
+    discoverUrl: `${MODELSCOPE_HOST}/mcp?hosted=1&page=1`,
+    apiKeyUrl: `${MODELSCOPE_HOST}/my/myaccesstoken`,
     tokenFieldName: 'modelScopeToken',
     getToken: getModelScopeToken,
     saveToken: saveModelScopeToken,
